brought changes from #252
