+++ src/pypromice/ressources/variable_aliases_GC-Net.csv
@@ -1,0 +1,78 @@
+GEUS_name,old_name
+time,timestamp
+p_u,P
+p_l,
+t_u,TA2
+t_l,TA1
+rh_u,RH2
+rh_u_cor,RH2_cor
+qh_u,Q2
+rh_l,RH1
+rh_l_cor,RH1_cor
+qh_l,Q1
+wspd_u,VW2
+wspd_l,VW1
+wdir_u,DW2
+wdir_l,DW1
+dsr,
+dsr_cor,ISWR
+usr,
+usr_cor,OSWR
+albedo,Alb
+dlr,
+ulr,
+cc,
+t_surf,
+dlhf_u,LHF
+dlhf_l,
+dshf_u,SHF
+dshf_l,
+z_boom_u,HW2
+z_boom_l,HW1
+precip_u,
+precip_u_cor,
+precip_l,
+precip_l_cor,
+t_i_1,TS1
+t_i_2,TS2
+t_i_3,TS3
+t_i_4,TS4
+t_i_5,TS5
+t_i_6,TS6
+t_i_7,TS7
+t_i_8,TS8
+t_i_9,TS9
+t_i_10,TS10
+t_i_11,
+tilt_x,
+tilt_y,
+rot,
+gps_lat,latitude
+gps_lon,longitude
+gps_alt,elevation
+gps_time,
+gps_geounit,
+gps_hdop,
+batt_v,V
+fan_dc_u,
+fan_dc_l,
+t_rad,
+msg_lat,
+msg_lon,
+z_surf_1,HS1
+z_surf_2,HS2
+z_surf_1_adj_flag,HS1_adj_flag
+z_surf_2_adj_flag,HS2_adj_flag
+z_surf_combined,HS_combined
+depth_t_i_1,DTS1
+depth_t_i_2,DTS2
+depth_t_i_3,DTS3
+depth_t_i_4,DTS4
+depth_t_i_5,DTS5
+depth_t_i_6,DTS6
+depth_t_i_7,DTS7
+depth_t_i_8,DTS8
+depth_t_i_9,DTS9
+depth_t_i_10,DTS10
+depth_t_i_11,
+t_i_10m,TS_10m
